chore(es-dev-server): rename etag cache
